refactor(twitter): extract tweet selection logic into SearchTweetSelector
- Moved tweet selection logic from TwitterSearchClient to a new SearchTweetSelector class to enhance modularity and adhere to the Single Responsibility Principle.
- Simplified the engageWithSearchTerms method by delegating tweet selection and validation responsibilities to the SearchTweetSelector.
- Improved code readability and maintainability by reducing complexity in the TwitterSearchClient.
